Move dashboard notice implementation to a new class
Creating a new `DashboardNotice` class
